refactor(web): AuthPage 拆出面板组件与纯格式化模块（984→424 行）
单文件塞了页面 + 7 个子面板 + 十余个纯函数。按「纯逻辑 / 组件 / 页面」三层拆开：

- auth-format.ts：状态归一（getPrimaryStatus 的 refresh_failed 分支）、百分比 / 金额 /
  窗口时长格式化、额度配色分档
- trend-chart-data.ts：趋势图类型、metric 常量、raw 点合并成图表行、坐标轴/tooltip 格式化
- UsageLimitsPanel.tsx：额度快照面板（Claude / Codex 两套卡片 + 新鲜度提示）
- UsageTrendPanel.tsx：剩余额度趋势图

JSX 逐字搬运，视觉零改动。纯模块与 metric-format.ts / inner-thought-format.ts 同惯例，
顺带补 18 个单测（此前这些逻辑埋在页面里没法测）。
